+++ jvm/src/main/scala/LemmatizedFormUrns.scala
@@ -1,0 +1,17 @@
+package edu.holycross.shot.latincorpus
+
+import edu.holycross.shot.ohco2._
+import edu.holycross.shot.cite._
+import edu.holycross.shot.tabulae._
+import edu.holycross.shot.mid.orthography._
+
+case class LemmatizedFormUrns(
+  psg: CtsUrn,
+  txt: String,
+  lexeme: Cite2Urn,
+  form: Cite2Urn
+) {
+  def cex(separator: String = "#") = {
+    List(psg, txt, lexeme, form).mkString(separator)
+  }
+}